backend/requirements.txt
flask
flask-cors
gunicorn
openai
requests
beautifulsoup4
python-dotenv
trafilatura